requirements.txt
Django==4.1.6
boto3-stubs==1.26.58
boto3==1.26.58
botocore-stubs==1.29.58
botocore==1.29.58
dj-database-url==1.0.0
django-allauth==0.51.0
django-ninja==0.20.0
email-validator==1.3.0
google-api-python-client-stubs==1.13.0
google-api-python-client==2.66.0
google-auth-oauthlib==0.7.1
gunicorn==20.1.0
humanize==4.4.0
psycopg2==2.9.5
whitenoise==6.3.0
django-auditlog==2.2.2
